message_store role does not own the schema
The message_store role does not need to own the message_store schema.
Allow it to be owned by the installing user, just like all of the other
objects.

We then grant the role `USAGE` privileges on the schema so that it can
access those objects.

This reduces the ability of the `message_store` role to execute DDL
to modify the schema's objects.

It also should address issues creating the schema in managed environments
like Amazon RDS. https://github.com/message-db/message-db/issues/1